move drag and drop context directly into criteria list component
- #330
